:recycle: 코드정리 및 테마코드 분리
